Reorder packages and room names:
renamed:    dojo-dungeon/src/dojo/rooms/level_3/R3_Fehler_Refactoring.java -> dojo-dungeon/src/dojo/rooms/rooms/boss/Fehler_Refactoring.java
renamed:    dojo-dungeon/src/dojo/rooms/level_1/R3_Fragen_Lambda.java -> dojo-dungeon/src/dojo/rooms/rooms/boss/Fragen_Lambda.java
renamed:    dojo-dungeon/src/dojo/rooms/level_4/R3_Fragen_RegExes.java -> dojo-dungeon/src/dojo/rooms/rooms/boss/Fragen_RegExes.java
renamed:    dojo-dungeon/src/dojo/rooms/level_2/R3_Monster_Implement_2.java -> dojo-dungeon/src/dojo/rooms/rooms/boss/Implement_MyImp.java
renamed:    dojo-dungeon/src/dojo/rooms/level_4/R2_Fehler_Quader.java -> dojo-dungeon/src/dojo/rooms/rooms/riddle/Fehler_Quader.java
renamed:    dojo-dungeon/src/dojo/rooms/level_1/R2_Fehler_Syntax.java -> dojo-dungeon/src/dojo/rooms/rooms/riddle/Fehler_Syntax.java
renamed:    dojo-dungeon/src/dojo/rooms/level_2/R2_Fragen_Pattern.java -> dojo-dungeon/src/dojo/rooms/rooms/riddle/Fragen_Pattern.java
renamed:    dojo-dungeon/src/dojo/rooms/level_3/R2_Monster_Implement_1.java -> dojo-dungeon/src/dojo/rooms/rooms/riddle/Implement_MyMonster.java
renamed:    dojo-dungeon/src/dojo/rooms/level_4/R1_Fragen_Schriftrollen.java -> dojo-dungeon/src/dojo/rooms/rooms/search/Fragen_Schriftrollen.java
renamed:    dojo-dungeon/src/dojo/rooms/level_1/R1_Monster_1.java -> dojo-dungeon/src/dojo/rooms/rooms/search/Key.java
renamed:    dojo-dungeon/src/dojo/rooms/level_2/R1_Monster_3.java -> dojo-dungeon/src/dojo/rooms/rooms/search/Monster_Kill.java
renamed:    dojo-dungeon/src/dojo/rooms/level_3/R1_Monster_2.java -> dojo-dungeon/src/dojo/rooms/rooms/search/Saphire.java

modified:   dojo-dungeon/src/dojo/rooms/builder/RoomBuilder.java
modified:   dojo-dungeon/src/starter/DojoStarter.java
