+++ dojo-dungeon/src/dojo/rooms/builder/RoomBuilder.java
@@ -6,18 +6,18 @@ import core.level.utils.LevelSize;
 import core.utils.components.path.IPath;
 import dojo.rooms.LevelRoom;
 import dojo.rooms.Room;
-import dojo.rooms.level_1.R1_Monster_1;
-import dojo.rooms.level_1.R2_Fehler_Syntax;
-import dojo.rooms.level_1.R3_Fragen_Lambda;
-import dojo.rooms.level_2.R1_Monster_3;
-import dojo.rooms.level_2.R2_Fragen_Pattern;
-import dojo.rooms.level_2.R3_Monster_Implement_2;
-import dojo.rooms.level_3.R1_Monster_2;
-import dojo.rooms.level_3.R2_Monster_Implement_1;
-import dojo.rooms.level_3.R3_Fehler_Refactoring;
-import dojo.rooms.level_4.R1_Fragen_Schriftrollen;
-import dojo.rooms.level_4.R2_Fehler_Quader;
-import dojo.rooms.level_4.R3_Fragen_RegExes;
+import dojo.rooms.rooms.boss.Fehler_Refactoring;
+import dojo.rooms.rooms.boss.Fragen_Lambda;
+import dojo.rooms.rooms.boss.Fragen_RegExes;
+import dojo.rooms.rooms.boss.Implement_MyImp;
+import dojo.rooms.rooms.riddle.Fehler_Quader;
+import dojo.rooms.rooms.riddle.Fehler_Syntax;
+import dojo.rooms.rooms.riddle.Fragen_Pattern;
+import dojo.rooms.rooms.riddle.Implement_MyMonster;
+import dojo.rooms.rooms.search.Fragen_Schriftrollen;
+import dojo.rooms.rooms.search.Key;
+import dojo.rooms.rooms.search.Monster_Kill;
+import dojo.rooms.rooms.search.Saphire;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -127,8 +127,8 @@ public class RoomBuilder {
    *
    * @return the newly generated room instance
    */
-  public Room buildRoom_Monster_1() {
-    return new R1_Monster_1(levelRoom, gen, nextRoom, levelSize, designLabel);
+  public Room buildRoom_Key() {
+    return new Key(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -137,7 +137,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fehler_Syntax() {
-    return new R2_Fehler_Syntax(levelRoom, gen, nextRoom, levelSize, designLabel);
+    return new Fehler_Syntax(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -146,7 +146,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fragen_Lambda() {
-    return new R3_Fragen_Lambda(levelRoom, gen, nextRoom, levelSize, designLabel);
+    return new Fragen_Lambda(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -154,8 +154,8 @@ public class RoomBuilder {
    *
    * @return the newly generated room instance
    */
-  public Room buildRoom_Monster_2() {
-    return new R1_Monster_2(levelRoom, gen, nextRoom, levelSize, designLabel);
+  public Room buildRoom_Saphire() {
+    return new Saphire(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -163,8 +163,8 @@ public class RoomBuilder {
    *
    * @return the newly generated room instance
    */
-  public Room buildRoom_Monster_Implement_1() {
-    return new R2_Monster_Implement_1(levelRoom, gen, nextRoom, levelSize, designLabel);
+  public Room buildRoom_Implement_MyMonster() {
+    return new Implement_MyMonster(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -173,7 +173,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fehler_Refactoring() {
-    return new R3_Fehler_Refactoring(levelRoom, gen, nextRoom, levelSize, designLabel);
+    return new Fehler_Refactoring(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -182,7 +182,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fragen_Schriftrollen() {
-    return new R1_Fragen_Schriftrollen(
+    return new Fragen_Schriftrollen(
         levelRoom, gen, nextRoom, levelSize, designLabel, monsterCount, monsterPaths, sortables);
   }
 
@@ -192,7 +192,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fehler_Quader() {
-    return new R2_Fehler_Quader(levelRoom, gen, nextRoom, levelSize, designLabel);
+    return new Fehler_Quader(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -201,7 +201,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fragen_RegExes() {
-    return new R3_Fragen_RegExes(levelRoom, gen, nextRoom, levelSize, designLabel);
+    return new Fragen_RegExes(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -209,8 +209,8 @@ public class RoomBuilder {
    *
    * @return the newly generated room instance
    */
-  public Room buildRoom_Monster_3() {
-    return new R1_Monster_3(levelRoom, gen, nextRoom, levelSize, designLabel);
+  public Room buildRoom_Monster_Kill() {
+    return new Monster_Kill(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -219,7 +219,7 @@ public class RoomBuilder {
    * @return the newly generated room instance
    */
   public Room buildRoom_Fragen_Pattern() {
-    return new R2_Fragen_Pattern(levelRoom, gen, nextRoom, levelSize, designLabel);
+    return new Fragen_Pattern(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 
   /**
@@ -227,7 +227,7 @@ public class RoomBuilder {
    *
    * @return the newly generated room instance
    */
-  public Room buildRoom_Monster_Implement_2() {
-    return new R3_Monster_Implement_2(levelRoom, gen, nextRoom, levelSize, designLabel);
+  public Room buildRoom_Implement_MyImp() {
+    return new Implement_MyImp(levelRoom, gen, nextRoom, levelSize, designLabel);
   }
 }
