+++ dojo-dungeon/src/dojo/rooms/rooms/boss/Fehler_Refactoring.java
@@ -1,4 +1,4 @@
-package dojo.rooms.level_3;
+package dojo.rooms.rooms.boss;
 
 import contrib.components.InteractionComponent;
 import contrib.hud.dialogs.OkDialog;
@@ -10,7 +10,6 @@ import core.level.utils.DesignLabel;
 import core.level.utils.LevelSize;
 import core.utils.IVoidFunction;
 import core.utils.components.path.SimpleIPath;
-import dojo.compiler.DojoCompiler;
 import dojo.rooms.LevelRoom;
 import dojo.rooms.Room;
 import dojo.rooms.TaskRoom;
@@ -21,14 +20,13 @@ import java.util.Set;
 /**
  * Informationen für den Spieler über diesen Raum:
  *
- * <p>In diesem Raum muss ein Monster mit verschiedenen Eigenschaften erstellt und danach besiegt
- * werden, um in den nächsten Raum zu gelangen.
+ * <p>In diesem Raum ist das Ziel, den vorgegebenen Code zu optimieren, Fehler zu beheben und die
+ * Lesbarkeit zu verbessern. Nur wenn der Code korrekt ist, kann man in den nächsten Raum
+ * weitergehen.
  */
-public class R2_Monster_Implement_1 extends TaskRoom {
-  private final String FILENAME1 = "../dojo-dungeon/todo-assets/Monster_Implement_1/Monster.java";
-  private final String FILENAME2 = "../dojo-dungeon/todo-assets/Monster_Implement_1/MyMonster.java";
-  private final String CLASS_NAME = "MyMonster";
-  private final String title = "Monster besiegen";
+public class Fehler_Refactoring extends TaskRoom {
+  private final String FILENAME1 = "ClsToRefactor.java";
+  private final String title = "Refactoring";
 
   /**
    * Generate a new room.
@@ -39,7 +37,7 @@ public class R2_Monster_Implement_1 extends TaskRoom {
    * @param levelSize the size of this room
    * @param designLabel the design label of this room
    */
-  public R2_Monster_Implement_1(
+  public Fehler_Refactoring(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
@@ -54,7 +52,7 @@ public class R2_Monster_Implement_1 extends TaskRoom {
           "Failed to generate: " + getClass().getName() + ": " + e.getMessage(), e);
     }
 
-    setRoomTitle("\"Tempel der verlorenen Geheimnisse\" (Raum 2)");
+    setRoomTitle("\"Tempel der verlorenen Geheimnisse\" (Raum 3)");
     setRoomDescription("Gehe zum Ritter für die Raumaufgabe.");
   }
 
@@ -63,31 +61,22 @@ public class R2_Monster_Implement_1 extends TaskRoom {
     IVoidFunction empty = () -> {};
     addTask(
         new Task(
-                this,
-                "task1",
-                empty,
-                () ->
-                    OkDialog.showOkDialog(
-                        String.format(
-                            "Implementiere die Datei %s, nach der Vorgabe in %s. Wenn das Monster besiegt ist, soll sich die Tür zum nächsten Raum öffnen.",
-                            FILENAME2, FILENAME1),
-                        title,
-                        empty),
-                (t1) -> {
-                  DojoCompiler.TestResult results =
-                      new DojoCompiler().spawnMonsterToOpenTheDoor(FILENAME2, CLASS_NAME, this);
-                  if (results.passed()) {
-                    OkDialog.showOkDialog(
-                        "Ok! " + results.messages(),
-                        title,
-                        () -> OkDialog.showOkDialog("Das Monster ist gespawnt!", title, empty));
-                    return true;
-                  }
-                  OkDialog.showOkDialog("Fehler: " + results.messages(), title, empty);
-                  return false;
-                },
-                empty)
-            .setShouldOpenDoors(false));
+            this,
+            "task1",
+            empty,
+            () ->
+                OkDialog.showOkDialog(
+                    String.format(
+                        "Refactor die Datei %s und lasse die korrigierte Datei von einem Tutor überprüfen.",
+                        FILENAME1),
+                    title,
+                    empty),
+            (t1) -> {
+              OkDialog.showOkDialog(
+                  "Das war schon alles, die Tür zum nächsten Level ist geöffnet.", title, empty);
+              return true;
+            },
+            empty));
 
     // Create questioner
     Entity questioner = new Entity();
@@ -109,7 +98,7 @@ public class R2_Monster_Implement_1 extends TaskRoom {
                         },
                         () ->
                             OkDialog.showOkDialog(
-                                "Das Monster ist bereits gespawnt!", title, empty))));
+                                "Die Aufgabe ist abgeschlossen!", title, empty))));
 
     // Add questioner to room
     addRoomEntities(Set.of(questioner));
