+++ dojo-dungeon/src/dojo/rooms/rooms/search/Saphire.java
@@ -1,4 +1,4 @@
-package dojo.rooms.level_1;
+package dojo.rooms.rooms.search;
 
 import contrib.level.generator.graphBased.RoomGenerator;
 import core.level.utils.DesignLabel;
@@ -12,18 +12,19 @@ import dojo.rooms.SearchRoom;
 /**
  * Informationen für den Spieler über diesen Raum:
  *
- * <p>In diesem Raum muss man den Schlüssel, der zufällig von einem Monster fallengelassen wird,
+ * <p>In diesem Raum muss man den Saphir, der zufällig von einem Monster fallengelassen wird,
  * finden, um in den nächsten Raum zu kommen.
  */
-public class R1_Monster_1 extends SearchRoom {
+public class Saphire extends SearchRoom {
   private static final int monsterCount = 5;
   private static final IPath[] monsterPaths = {
-    new SimpleIPath("character/monster/imp"), new SimpleIPath("character/monster/goblin")
+    new SimpleIPath("character/monster/orc_shaman"),
+    new SimpleIPath("character/monster/orc_warrior")
   };
 
-  private static final String keyType = "Golden Key";
-  private static final String keyDescription = "A key to unlock the next room.";
-  private static final IPath keyTexture = new SimpleIPath("items/key/gold_key.png");
+  private static final String keyType = "A blue gemstone";
+  private static final String keyDescription = "This gem opens the door to the next room.";
+  private static final IPath keyTexture = new SimpleIPath("items/resource/saphire.png");
 
   /**
    * Generate a new room.
@@ -34,7 +35,7 @@ public class R1_Monster_1 extends SearchRoom {
    * @param levelSize the size of this room
    * @param designLabel the design label of this room
    */
-  public R1_Monster_1(
+  public Saphire(
       LevelRoom levelRoom,
       RoomGenerator gen,
       Room nextRoom,
@@ -52,8 +53,8 @@ public class R1_Monster_1 extends SearchRoom {
         keyDescription,
         keyTexture);
 
-    setRoomTitle("\"Der vergessene Wald\" (Raum 1)");
+    setRoomTitle("\"Tempel der verlorenen Geheimnisse\" (Raum 1)");
     setRoomDescription(
-        "Du bist in einem Raum voller Monster. Besiege die Monster und finde den Schlüssel für den nächsten Raum.");
+        "Du bist in einem Raum voller Monster. Besiege die Monster und finde den Saphir für den nächsten Raum.");
   }
 }
